+++ buildSrc/src/main/groovy/org/gradle/GroovyCustomTask.groovy
@@ -4,7 +4,7 @@ import org.gradle.api.DefaultTask
 import org.gradle.api.tasks.Input
 import org.gradle.api.tasks.TaskAction
 
-class CustomTask extends DefaultTask {
+class GroovyCustomTask extends DefaultTask {
   @Input
   String greeting = 'Hello from within the GreetingTask'