+++ src/test/groovy/org/gradle/GroovyCustomTaskTest.groovy
@@ -1,0 +1,8 @@
+class GroovyCustomTaskTest {
+  @Test
+  public void canAddTaskToProject() {
+    Project project = ProjectBuilder.builder().build()
+    def task = project.task('greeting', type: GroovyCustomTask)
+    assertTrue(task instanceof GroovyCustomTask)
+  }
+}