+++ CHANGELOG.md
@@ -5,3 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## 2019-12-01 [0.0.2]
+### Added
+- Gradle Wrapper
+- Added buildSrc/ Java and Groovy files.